requirements.txt
groq
spacy
streamlit
duckduckgo-search
requests
beautifulsoup4
python-dotenv
